src/echomind_lib/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "echomind-lib"
version = "0.1.0"
description = "EchoMind Shared Library"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "EchoMind Team"}
]

# Shared dependencies - ALL services must use these exact versions
dependencies = [
    # Core
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "python-dotenv==1.2.1",
    # Database
    "sqlalchemy[asyncio]==2.0.45",
    "asyncpg==0.31.0",
    # Vector Database
    "qdrant-client==1.16.2",
    # Cache
    "redis==7.1.0",
    # Object Storage
    "miniopy-async==1.23.4",
    # Message Queue
    "nats-py==2.12.0",
    # Authentication
    "PyJWT==2.10.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "aiosqlite==0.21.0",
    "httpx==0.28.1",
    # Linting & Formatting
    "black==26.1.0",
    "ruff==0.14.13",
    "mypy==1.19.1",
    # Proto Generation
    "grpcio-tools==1.69.0",
    "protobuf==5.29.5",
    "mypy-protobuf==3.6.0",
    "protobuf-pydantic-gen==0.1.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["echomind_lib*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
